moved usim utilities to __init__ of testing module
